[rANS] Aggregated metrics to compute estimates
`DatasetMetrics` are computed from a distribution of source symbols
provided by a frequency table. The distribution is used to compute
entropy and derive the distribution of optimal symbol lengths,
based on the frequency of a source symbol. A distribution of
optimal symbol lengths weighted by the frequency of each symbol
is used to estimate a renorming weight and a cut-off point,that
decides determins which symbols should be replaced by an
incompressible symbol during renorming