[rANS] `BitPtr` abstraction
Create pointer abstraction that allows pointing to individual Bits in
contiguos memory efficiently.

BitPtr uses the fact that X86 and aarch64 systems don't use the full 64
Bit address space but only 48 to 52 bits, which allows us to use a conversion to `intptr_t` to
address individual bits. Requires all basic types to be aligned as
defined by the C++ standard. This will not be portable to systems that
use more than 58 Bits of physical + virtual address space (>288 Pb of
RAM).